ch08/requirements.txt
amqp==5.1.1
anyio==3.6.1
asyncpg==0.26.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
fastapi==0.78.0
flower==1.1.0
gino==1.0.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
humanize==4.2.3
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
kafka-python==2.0.2
kombu==5.2.4
priority==2.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.1
python-multipart==0.0.5
pytz==2022.1
rfc3986==1.5.0
Rx==3.2.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
sse-starlette==0.10.3
starlette==0.19.1
toml==0.10.2
tornado==6.2
typing_extensions==4.3.0
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
websockets==10.3
wsproto==1.1.0
